UserSecurityFunc klasse gemaakt voor user security checks. userId verwijdert uit HouseDTO security checks toegepast op de methods in de controllers methode aan House_UserDAO toegevoegd

+++ SmartHome/src/main/java/com/ucll/smarthome/controller/HouseController.java
@@ -1,13 +1,14 @@
 package com.ucll.smarthome.controller;
 
 import com.ucll.smarthome.dto.HouseDTO;
-import com.ucll.smarthome.dto.UserDTO;
+import com.ucll.smarthome.functions.UserSecurityFunc;
 import com.ucll.smarthome.persistence.entities.House;
 import com.ucll.smarthome.persistence.entities.User;
 import com.ucll.smarthome.persistence.repository.HouseDAO;
 import com.ucll.smarthome.persistence.repository.UserDAO;
 import com.vaadin.flow.router.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AuthorizationServiceException;
 import org.springframework.stereotype.Controller;
 
 import javax.transaction.Transactional;
@@ -22,13 +23,15 @@ public class HouseController {
     private final HouseDAO dao;
     private final UserDAO userDAO;
     private final House_UserController house_userController;
+    private final UserSecurityFunc userSecurityFunc;
 
 
     @Autowired
-    public HouseController(HouseDAO dao, UserDAO userDAO, House_UserController house_userController) {
+    public HouseController(HouseDAO dao, UserDAO userDAO, House_UserController house_userController, UserSecurityFunc userSecurityFunc) {
         this.dao = dao;
         this.userDAO = userDAO;
         this.house_userController = house_userController;
+        this.userSecurityFunc = userSecurityFunc;
     }
 
     /**
@@ -39,9 +42,10 @@ public class HouseController {
     public void createHouse(HouseDTO houseDTO) throws IllegalArgumentException{
         if (houseDTO == null) throw new IllegalArgumentException("Creating house failed. Iputdata missing.");
         if (houseDTO.getName() == null || houseDTO.getName().length() == 0) throw new IllegalArgumentException("Creating house failed. Name of the house not filled in.");
-        if (houseDTO.getUserid() <= 0L) throw new IllegalArgumentException("Can't create without a user id");
 
-        Optional<User> user = userDAO.findById(houseDTO.getUserid());
+        long userId = userSecurityFunc.getLoggedInUserId();
+
+        Optional<User> user = userDAO.findById(userId);
         if (user.isEmpty()) throw new IllegalArgumentException("User does not exist");
 
         House house = houseDtoToHouse(houseDTO);
@@ -61,13 +65,18 @@ public class HouseController {
         if (houseDTO == null) throw new IllegalArgumentException("Updating house failed. Iputdata missing.");
         if (houseDTO.getName() == null || houseDTO.getName().trim().equals("")) throw new IllegalArgumentException("Updatin house failed. Name of the house not filled in.");
 
+        long userId = userSecurityFunc.getLoggedInUserId();
 
+        Optional<House> h = dao.findById(userId);
 
-        Optional<House> h = dao.findById(houseDTO.getId());
-        if (h.isEmpty()) {
-            throw new NotFoundException("The house you want to update is not found");
+        if (h.isPresent()) {
+            if(userSecurityFunc.checkCurrentUserIsAdmin(h.get().getHouseId())){
+                h.get().setName(houseDTO.getName());
+            }else{
+                throw new AuthorizationServiceException("You are not an admin of this house.");
+            }
         } else {
-            h.get().setName(houseDTO.getName());
+            throw new NotFoundException("The house you want to update is not found.");
         }
 
     }
@@ -81,22 +90,23 @@ public class HouseController {
     public HouseDTO getHouseById(long houseId) throws IllegalArgumentException{
         if (houseId <= 0L) throw new IllegalArgumentException("House Id is missing");
         Optional<House> house = dao.findById(houseId);
-        if (house.isPresent()){
-            return new HouseDTO.Builder().id(house.get().getHouseId()).name(house.get().getName()).build();
-        }else {
-            throw new IllegalArgumentException("No house found with id: " + houseId);
-        }
+
+        if (house.isEmpty()) throw new IllegalArgumentException("No house found with id: " + houseId);
+        if (userSecurityFunc.getHouseUser(houseId).isEmpty()) throw new IllegalArgumentException("No house found with id: " + houseId);
+
+        return new HouseDTO.Builder().id(house.get().getHouseId()).name(house.get().getName()).build();
     }
 
     /**
      * find the houses by a user
-     * @param userid id to search for a user
      * @return list of housedto
      * @throws IllegalArgumentException if something goes wrong with the info what went wrong
      */
-    public List<HouseDTO> getHousesByUser(long userid) throws IllegalArgumentException{
-        if (userid <= 0L) throw new IllegalArgumentException("Invalid id");
-        Optional<User> u = userDAO.findById(userid) ;
+    public List<HouseDTO> getHousesByUser() throws IllegalArgumentException{
+
+        long userId = userSecurityFunc.getLoggedInUserId();
+
+        Optional<User> u = userDAO.findById(userId) ;
         if (u.isEmpty()){
             throw new IllegalArgumentException("User not found to get houses");
         }else{
@@ -104,7 +114,6 @@ public class HouseController {
                     .map(rec -> new HouseDTO.Builder()
                             .id(rec.getHouse().getHouseId())
                             .name(rec.getHouse().getName())
-                            .userid(rec.getUser().getId())
                             .build());
             return stream.collect(Collectors.toList());
         }
@@ -130,9 +139,7 @@ public class HouseController {
     public void deleteHouse(long houseid) throws IllegalArgumentException {
         if (houseid <= 0L) throw new IllegalArgumentException("Invalid id");
 
-        //checking if the house exists
-        getHouseById(houseid);
-
+        if(!userSecurityFunc.checkCurrentUserIsAdmin(houseid))  throw new IllegalArgumentException("The house you want to delete is not yours");
 
         house_userController.deleteRegistratieHouseUser(house_userController.getByHouse(dao.getById(houseid)));
         dao.deleteById(houseid);
